fixed warnings and added tare functions

+++ src/main/cpp/str/SwerveDriveSim.cpp
@@ -12,21 +12,21 @@ using namespace str;
 
 void SwerveDriveSim::SetDriveInputs(const std::array<units::volt_t, 4>& inputs)
 {
-  for (int i = 0; i < simModules.size(); i++) {
+  for (size_t i = 0; i < simModules.size(); i++) {
     simModules[i].driveMotor.SetInputVoltage(inputs[i]);
   }
 }
 
 void SwerveDriveSim::SetSteerInputs(const std::array<units::volt_t, 4>& inputs)
 {
-  for (int i = 0; i < simModules.size(); i++) {
+  for (size_t i = 0; i < simModules.size(); i++) {
     simModules[i].steerMotor.SetInputVoltage(inputs[i]);
   }
 }
 
 void SwerveDriveSim::Update(units::second_t dt)
 {
-  for (int i = 0; i < simModules.size(); i++) {
+  for (size_t i = 0; i < simModules.size(); i++) {
     simModules[i].driveMotor.Update(dt);
     simModules[i].steerMotor.Update(dt);
   }
@@ -35,7 +35,7 @@ void SwerveDriveSim::Update(units::second_t dt)
 units::ampere_t SwerveDriveSim::GetDriveCurrentDraw() const
 {
   units::ampere_t total;
-  for (int i = 0; i < simModules.size(); i++) {
+  for (size_t i = 0; i < simModules.size(); i++) {
     total = total + simModules[i].driveMotor.GetCurrentDraw();
   }
   return total;
@@ -44,7 +44,7 @@ units::ampere_t SwerveDriveSim::GetDriveCurrentDraw() const
 units::ampere_t SwerveDriveSim::GetSteerCurrentDraw() const
 {
   units::ampere_t total;
-  for (int i = 0; i < simModules.size(); i++) {
+  for (size_t i = 0; i < simModules.size(); i++) {
     total = total + simModules[i].steerMotor.GetCurrentDraw();
   }
   return total;
@@ -54,7 +54,7 @@ std::array<SimState, 4> SwerveDriveSim::GetState() const
 {
   std::array<SimState, 4> state;
 
-  for (int i = 0; i < simModules.size(); i++) {
+  for (size_t i = 0; i < simModules.size(); i++) {
     state[i].drivePos = SwerveModule::ConvertMotorToWheelDistance(
       simModules[i].driveMotor.GetAngularPosition()
       * constants::swerve::physical::DRIVE_GEARING);
